+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # ArCode versions
 
+## 1.3.2
+- Add ArCode Gallery section.
+- Load user profile from arweave-account.
+
 ## 1.3.1
 Link (2022-10-19): https://arweave.net/8n2ZJ4EzNRZg8XfnSx75VAs_PlHBfElsj7CKfPC9mYw
-- Fix links
+- Fix links.
 
 ## 1.3.0
 Link (2022-10-19): https://arweave.net/HRww-Zu2bZiFZZuCYo0fN1e9KWcvTDCP8xW55bkEvwk